Refactor controller structure
