+++ pkg/controllers/endpointmonitor/endpointmonitor_created.go
@@ -1,16 +1,17 @@
-package endpointmonitor
+package controllers
 
 import (
-	endpointmonitorv1alpha1 "github.com/stakater/IngressMonitorController/pkg/apis/endpointmonitor/v1alpha1"
+	endpointmonitorv1alpha1 "github.com/stakater/IngressMonitorController/api/v1alpha1"
 	"github.com/stakater/IngressMonitorController/pkg/kube/util"
 	"github.com/stakater/IngressMonitorController/pkg/models"
 	"github.com/stakater/IngressMonitorController/pkg/monitors"
 
-	log "github.com/sirupsen/logrus"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
-func (r *ReconcileEndpointMonitor) handleCreate(request reconcile.Request, instance *endpointmonitorv1alpha1.EndpointMonitor, monitorName string, monitorService monitors.MonitorServiceProxy) error {
+func (r *EndpointMonitorReconciler) handleCreate(request reconcile.Request, instance *endpointmonitorv1alpha1.EndpointMonitor, monitorName string, monitorService monitors.MonitorServiceProxy) error {
+	log := r.Log.WithValues("endpointMonitor", instance.ObjectMeta.Namespace)
+
 	log.Info("Creating Monitor: " + monitorName)
 
 	url, err := util.GetMonitorURL(r.client, instance)
